Correcciones en las clases Notas! * Se cambio a int la variable de frecuencia para requerir menor uso de memoria * Debido a que se requirio cambiar los valores de frecuencias, corregui los test * Ahora corren el 100% los test

+++ AlgoHero/src/algo3c1g2/modelo/nota/LaSostenido.java
@@ -5,7 +5,7 @@ import algo3c1g2.modelo.figura.Figura;
 public class LaSostenido extends Nota {
 
 	//Frecuencia caracteristica del tono La Sostenido en la octava 3
-	private final static float FRECUENCIA_LA_SOSTENIDO = 70;
+	private final static int FRECUENCIA_LA_SOSTENIDO = 70;
 	private final static int CUERDA = 6;
 	
 	public LaSostenido(){
