+++ AlgoHero/src/algo3c1g2/modelo/nota/ReSostenido.java
@@ -5,7 +5,7 @@ import algo3c1g2.modelo.figura.Figura;
 public class ReSostenido extends Nota {
 
 	//Frecuencia caracteristica del tono Re Sostenido en la octava 3
-	private final static float FRECUENCIA_RE_SOSTENIDO = 63;
+	private final static int FRECUENCIA_RE_SOSTENIDO = 63;
 	private final static int CUERDA = 2;
 	
 	public ReSostenido(){
