+++ AlgoHero/src/algo3c1g2/modelo/nota/Si.java
@@ -5,7 +5,7 @@ import algo3c1g2.modelo.figura.Figura;
 public class Si extends Nota {
 
 	//Frecuencia caracteristica del tono Si en la octava 3
-	private final static float FRECUENCIA_SI = 71;
+	private final static int FRECUENCIA_SI = 71;
 	private final static int CUERDA = 6;
 	
 	public Si(){
